Moved Bot to its own file
